Converted patch elements types to int from float.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@2075 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/connections/SubunitConn.cpp
@@ -44,13 +44,13 @@ int SubunitConn::weights(PVPatch * wp)
 {
    pvdata_t * w = wp->data;
 
-   const int nx = (int) wp->nx;
-   const int ny = (int) wp->ny;
-   const int nf = (int) wp->nf;
+   const int nx = wp->nx;
+   const int ny = wp->ny;
+   const int nf = wp->nf;
 
-   const int sx = (int) wp->sx;  assert(sx == nf);
-   const int sy = (int) wp->sy;  assert(sy == nf*nx);
-   const int sf = (int) wp->sf;  assert(sf == 1);
+   const int sx = wp->sx;  assert(sx == nf);
+   const int sy = wp->sy;  assert(sy == nf*nx);
+   const int sf = wp->sf;  assert(sf == 1);
 
    assert(nx == 3);
    assert(ny == 3);
